Fix a bunch of warnings
using the list of warnings from here:
https://github.com/lefticus/cppbestpractices/blob/e73393f25a85f83fed7399d8b65cb117d00b2231/02-Use_the_Tools_Available.md#L100

Seems like a good place to start, and would improve things a bit
type-wise.  This patchset attempts to correct all the issues in one
shot.

Tested:
It builds.  Will test various subsystems that have been touched

+++ redfish-core/include/registries.hpp
@@ -34,7 +34,7 @@ struct Message
     const char* description;
     const char* message;
     const char* severity;
-    const int numberOfArgs;
+    const size_t numberOfArgs;
     std::array<const char*, 5> paramTypes;
     const char* resolution;
 };
